requirements.txt
async-timeout==4.0.3
blinker==1.8.2
boto3==1.35.49
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gunicorn==23.0.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
peewee==3.17.6
pyjwt==2.9.0
redis==5.0.8
requests==2.32.3
rq==1.16.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2